Provide an API for configuring TLS client authentication
This commit adds a simple API for specifying the TLS certificate a GRPC
client will present (via the same `Identity` wrapper as a server cert is
configured). It also adds an API to specify which CA certificate client
TLS certificates will be validated against for servers.

To demonstrate usage, a new example `tls_client_auth` is added, using
OpenSSL for the client and Rustls for the server.