feat(build): Reflection support
This commit adds a new option, `include_file_descriptor_set` to the
tonic build configuration to in order to include the encoded protocol
buffers `FileDescriptorSet` in `OUT_DIR`.

It also adds a new macro `include_file_descriptor_set!`, to make it easy
to include the well-known output path as a byte array in application
source code.