Endpoint::lazy_connect always tries to reconnect
When a connection could not be established on first connection
attempt, an error was reported and it never tried to  reconnect.
This behaviour is kepts for `Endpoint::connect` and changed for
`Endpoint::lazy_connect`.

Rationale: When connecting lazily, we expect the connection to
be always resilient, since we move the actual connection to a later
point in time. This is also useful, when connecting to services
that are not yet started.