Update Unix Domain Socket example to provide concrete type
In the process of updating our code to tokio 1.x, I had a little more
trouble migrating the code I had on tokio 0.2 because the types that we
used disappeared, particularly tokio::next::unit::Incoming.  Drawing on
the example of hyper::server::conn::AddrIncoming, this creates a
UnixIncoming struct in the example, and then uses that within the server
code.

I generally found this cleaner, though I understand that is subjective.
I also thought it could be useful to inform any further unix api
discussions, similar in the vein of #365.