feat(tonic): pass `trace_fn` the request rather than just the headers
Previously `Server::trace_fn` would receive `&HeaderMap` which meant you
couldn't add things like the URI to the span.

With this change it now receives `&http::Request<()>`. It doesn't
receive the body since that requires an additional generic parameter
which spreads to a lot of places. There also isn't much you can do with
the body of a borrowed `http::Request` anyways so its probably fine.