feat: Update `prost` and friends` to 0.9
This commit updates `prost`, `prost-build`, and `prost-types` to v0.9
for all `tonic` crates.

Since updating the `prost` version is a breaking change, I've also gone
ahead and bumped all the crate versions to 0.6. I hope that was correct
to do as part of this branch, but I can back it out if not.

BREAKING CHANGE: `prost` bumped to `0.9`.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>