chore: Streaming<T> reduce monomorphization cost
This was seen using `cargo llvm-lines`.

The following functions where high in generated lines with a few copies
(~ one copy by send message type).

- <tonic::codec::decode::Streaming<T> as futures_core::stream::Stream>::poll_next
- tonic::codec::decode::Streaming<T>::decode_chunk

Both of them contains a large part of code that is actually not impacted
by the type parameter and so can be reworked a bit to improve code
generation (should help produce small binaries and reduce build time).