Server signals ephemeral port when ready
Creating client/server unit tests has two classical problems:

1) How do you find a port that's unused that the server can
    bind but also communicate it to the client.  Typically
    hard coding one in a test means the test will fail anytime
    that port is bound (1338 in current code)
2) How do you start the server and the client as fast as possible
    (think 100s of tests running) without using a sleep() which
    either is too small (causing race conditions) or too large
    (causing testing delays)

This patch proposes a way for the client to register a signal
and the server to send the ephemeral port back on that signal
when it's ready, solving both problems.

I created an example unit test to show how this could be used
but I can fix up the rest and add the right documentation if
folks agree with the general direction.

Feedback welcome.