Decouple `NamedService` from the `transport` feature
The `tonic::transport::NamedService` trait is gated by the `transport`
feature. This means that generated server implementations can only
include this implementation when the `transport` feature is enabled on
`tonic-build`; and if the `transport` feature is not used at
generation-time, then a server cannot be used with
`tonic::transport::Server` at runtime.

We would like to publish a crate including generated protobuf/gRPC
bindings that does not _require_ a dependency on the `transport` feature
(and its fairly heavyweight dependency tree). But it would be nice if
its servers could still be used by applications that want to use
`tonic::transport::Server`.

By decoupling the `NamedService` trait from the `transport` feature,
`tonic-build` can be changed to always provide an implementation for
this trait without requiring all of the `transport` dependencies. This
enables generated servers to be used with `tonic::transport::Server`.