tonic-build: add module for declaratively generating services
This patch adds the `rust` module to tonic-build which provides
utilities for declaratively defining services in rust code instead of
via the traditional method of defining a service in a `proto` file.

It also enables setting a custom `Codec` to in order to use a custom
serialization format other than `protobuf`.