tonic-build: Add option to emit cargo:rerun-if-changed instructions
Currently, tonic-build does not emit cargo:rerun-if-changed
instructions. This will cause the build script to be rerun if any file
changes in the package.

This commit implements an option on the Builder called
emit_rerun_if_changed that, when enabled, will emit the
cargo:rerun-if-changed instructions for all of the proto file paths
and include directories.

Fixes: #1020
Refs: #1019