feat: support the SSLKEYLOGFILE environment variable for clients
Nothing changes unless SSLKEYLOGFILE is set in the environment.

If SSLKEYLOGFILE is set, then every time the client connnects, it will
append a line to the file designated in SSLKEYLOGFILE with the pre-master
secret for that connection.

Point a tool like Wireshark to that file, and suddenly it can decode
the TLS conversations.

https://wiki.wireshark.org/TLS#using-the-pre-master-secret

https://docs.rs/rustls/0.20.6/rustls/struct.KeyLogFile.html

https://docs.rs/rustls/0.20.6/rustls/trait.KeyLog.html