fix(web): Buffer incomplete messages
Previously, if we did not find trailers in the incoming message and the
message len was longer than the actual buffer we returned the buffer.
This would mean we would fail to parse incomplete message buffers. This
fixes `find_trailers` to support three return types, `Trailers(len)`,
`IncompleteBuf`, `Done`. Previously, there was only `Some(len)` and
`None`.