Expose prost-build's skip_debug method via configure
When generating types from protobuf messages, it can be useful to
manually implement Debug (eg to skip or alter some fields). This has
been enabled in recent prost-build versions and this extends that to a
method on the configuration builder.

It was previously possible to call the skip_debug method by creating an
instance of Config and passing it to compile_with_config so this is
purely a convenience method.

    let mut conf = Config::new();
    conf.skip_debug("TypeToSkip");
    tonic_build::configure()
        .compile_with_config(conf, &[...], &[...])?;

becomes

    tonic_build::configure()
        .skip_debug("TypeToSkip")
        .compile(&[...], &[...])?;

This also means that client code no longer needs to include a
dev-dependency on prost-build to use this feature.