Convert from hyper::Body to http_body::BoxedBody
When appropriate, we replace `hyper::Body` with `http_body::BoxedBody`, a good general purpose replacement for `hyper::Body`.

Hyper does provide `hyper::body::Incoming`, but we cannot construct that, so anywhere we might need a body that we can construct (even most Service trait impls) we must use something like `http_body::BoxedBody`.

When a service accepts `BoxedBody` and not `Incoming`, this indicates that the service is designed to run in places where it is not adjacent to hyper, for example, after routing (which is managed by Axum)

Additionally, http >= 1 requires that extension types are `Clone`, so this bound has been added where appropriate.

Co-authored-by: Ivan Krivosheev <py.krivosheev@gmail.com>
Co-authored-by: Allan Zhang <allanzhang7@gmail.com>