streaming: terminate stream after an error has been returned once (#1823)
PR #1682 introduced a change to return the error Status to the user of
Streaming but resulted in a change in behavior such that if an error
occurs the Stream will never end and instead always yeild the error.
This patch adjusts the semantics to only return the error one time and
then terminates the stream by returning Poll::Ready(None) on subseqent
calls.