feat: Add tcp setting getters to endpoint (#1983)
Add getters for tcp settings to Endpoints. The purpose for this is that
`connect_with_connector` doesn't set TCP settings on the socket, because
it's using a provided connector. In the cases where a library is creating
a custom TCP connection, we want to be able to respect user tcp
settings, in particular connection timeout. Adding these getters allows
this.