examples: switch version to last released version (#2047)
As part of https://github.com/hyperium/tonic/pull/2014, the versions of
"tonic" and "tonic-build" were changed to "0.13". This breaks the
instuctions in the examples documentation with an error like:
```
helloworld_tonic$ cargo run --bin helloworld-server
    Updating crates.io index
error: failed to select a version for the requirement `tonic = "^0.13"`
candidate versions found which didn't match: 0.12.3, 0.12.2, 0.12.1, ...
location searched: crates.io index
required by package `helloworld-tonic v0.1.0 (/usr/local/google/home/easwars/src/rust/projects/helloworld_tonic)`
if you are looking for the prerelease package it needs to be specified explicitly
    tonic = { version = "0.1.0-beta.1" }
```

This PR changes the version of "tonic" and "tonic-build" to "*" which
results in the latest version getting picked. Setting it to "*" also has
the nice property of not having to update the examples everytime a
version bump is required.

Co-authored-by: Lucio Franco <luciofranco14@gmail.com>