feat(build): Add 'use_generic_streaming_requests'
Adds an option for server trait generation that causes client-streaming
methods to receive a generic `impl IntoStreamingRequest` param
instead of a concrete `Request<Streaming<...>>` param.
This allows for testing trait implementions with any type implementing `Stream`,
and not having to go through the trouble of encoding a `Streaming`
object.

Fixes #462