feat(tonic): rpit future for generated server traits
Supported at 1.75, which is tonic's current MSRV, return position
impl trait (RPIT) makes user code cleaner, friendlier in a debugger,
and possibly a couple nanoseconds faster due to fewer boxed Future
indirections. There is still a Box Future at the tower_service interface
due to the switch from RPIT style "simple" traits and the fancier static
associated type Future style from tower.

It's similar to https://github.com/hyperium/tonic/pull/1697 but instead
I went for the full change, to evaluate the expected impact more fully.

In the expected use cases where people generate their code, they would
remove `#[tonic::async_trait]` from their implementations of Server
stubs. You can see in all the tests and examples, that's the only
user-facing delta.

It is a breaking change, but it's a simplification that seemed
worthwhile to me with the burden of 1-line deletion per service.
The vs code auto-generated todo!() implementations are a little nicer
this way too, which is a quality of life boost for me :-)

Relates to: https://github.com/hyperium/tonic/issues/1651

________________
**UX before:**
```rust
 #[tonic::async_trait]
 impl Greeter for MyGreeter {
```

**UX after:**
```rust
 impl Greeter for MyGreeter {
```

________________

**Testing**

Due to the nature of the change, a wide range of tests exercise it.
It is a change to the existing infrastructure rather than a new
feature, so there's no duplication or parameterization of tests
that would increase coverage.