Update helloworld-tutorial.md (#2385)
Update the hello-world tutorial readme to reflect changes to package
names and remove reference to deprecated tools.

## Motivation

As someone relatively new to Rust and new to Tonic, I was working
through the hello-world tutorial and found a few issues. I've
successfully built the project with the changes in this PR.

## Solution

- update example Cargo.toml to reflect package name updates
- update example build.rs with correct build tool name
- remove reference to deprecated GUI tool

I've tested this and it runs and builds correctly.

I sort of hate to reference Postman as the recommended GUI but it was
the first one on the list referenced by the now-deprecated Bloom GUI.