fix(prost-build): Map wrapper types to primitives
This commit address the issue raised in #2388, where well-known types
such as `google.protobuf.BoolValue` would appear as `super::bool` rather
than the correct primitives (`bool` in this case).

We address this by adding new cases to the existing enum for the various
wrapper types contained in `google/protobuf/wrappers.proto`, when not
compiling well-known types as proto structs directly.